+++ odoo/local-src/specific_base/models/__init__.py
@@ -9,6 +9,7 @@ from . import calendar_event
 from . import pricelist
 from . import account_invoice
 from . import res_users
+from . import crm_activity
 from . import crm_lead
 from . import crm_stage
 from . import utm