+++ HISTORY.rst
@@ -11,6 +11,7 @@ Unreleased
 * DT-78: Update all translations for specific modules
 * DT-86: Change precision of latitude and longitude fields on Company
 * DT-86: Change type of capital stock field from Integer to Float on Company
+* DT-87: Display action date in crm lead like a datetime
 * DT-91: On views 'My Pipeline' and 'Next Activities', remove default filter